backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
cryptography==41.0.7
email-validator==2.1.0

# Rate limiting
slowapi==0.1.9

# Environment & Configuration
python-dotenv==1.0.0

# HTTP clients
httpx==0.25.2
requests==2.31.0

# Logging & Monitoring
python-json-logger==2.0.7
psutil==5.9.6